refactor naming and split xferdtensor kernel into its own module
Naming improvements (no behavior change):
- normalize_refit_info_placements → restore_refit_info_placements
  (the helper undoes msgspec's lossy dict-serialization; "restore" /
  "rehydrate" is more accurate than "normalize" / "canonicalize")
- _normalize_placement → _restore_placement
- finalize closure → post_refit_hook on the gen-side merged-param path
- dim0_slice → merged_param_slice (the slice now reads as a signal that
  the param is part of a merge, not as a mechanical "slice on dim 0")
- _meta_metadata_context → _meta_tensor_alloc_context (the double "meta"
  was confusing — "meta" device vs tensor "metadata")
- _moe_kind → fused_expert_param_type (now scoped enough to be precise;
  also renamed the matching function parameter `kind` → same)

Module split:
- New nemo_rl/distributed/xferdtensor.py owns the transfer kernel:
  the four private helpers (_flatten_mesh_ranks, _get_tensor_meta,
  _get_mesh_coords, _compute_shard_slices), xferdtensor_golden (the
  broadcast-based reference), and a public xferdtensor() wrapper that
  currently forwards to the golden impl.  Replacing xferdtensor()'s
  body with the real low-level NCCL-XferReshard API will be a one-file
  change with no caller updates needed.
- nccl_reshard_utils.py keeps the wrapper types (MeshInfo, DTensorRef)
  and the metadata builders (placement rules, MoE fusion,
  build_nccl_reshard_refit_info, restore_refit_info_placements).  The
  kernel reads these by duck typing so xferdtensor.py has zero project-
  internal imports.
- Dead code removed: TensorWrapper (defined but never imported).
- All callers (dtensor_policy_worker_v2, megatron_policy_worker,
  vllm_backend) now import and call xferdtensor (not _golden).
  Higher-level files (grpo, lm_policy, vllm_worker, vllm_generation)
  had textual references updated for consistency.

Verified on the full L1 matrix (11 L0 + dsv3 + qwen3-235b).  12/13 jobs
green with KL matching baselines.  dsv3 crashed in an unrelated
sequence-packing path inside the loss code (_pack_input_ids,
nemo_rl/algorithms/loss/utils.py:208) after producing 7 healthy refit
KL values (0.0018–0.0023); pre-existing async-GRPO flakiness, not a
refactor regression.
