+++ nemo_rl/models/policy/lm_policy.py
@@ -965,7 +965,7 @@ class Policy(ColocatablePolicyInterface, GenerationInterface):
         return results[0]
 
     def nccl_reshard_refit(self, kv_scales=None) -> list[ray.ObjectRef]:
-        """Transfer weights to gen workers via nccl_reshard (xferdtensor_golden)."""
+        """Transfer weights to gen workers via nccl_reshard (xferdtensor)."""
         futures = self.worker_group.run_all_workers_single_data(
             "nccl_reshard_refit",
             kv_scales=kv_scales,
